requirements.txt
django==2.1.2
libsass==0.15.1
django-compressor==2.2
django-sass-processor==0.7.1
mysqlclient==1.3.13
fabric3==1.14.post1
fabistrano==0.3